+++ lib/restart-trigger.ts
@@ -1,2 +1,2 @@
-// restart triggered at 2026-03-24T14:23:49.520Z
-export const RESTART_TS = 1774362229520;
+// restart triggered at 2026-03-25T07:29:29.019Z
+export const RESTART_TS = 1774423769019;